feat: [IDP-0]: Fix package names
