+++ 101-idp-service/src/main/java/io/harness/dtos/AppConfigDTO.java
@@ -1,0 +1,12 @@
+package io.harness.dtos;
+
+import io.harness.annotations.dev.HarnessTeam;
+import io.harness.annotations.dev.OwnedBy;
+import lombok.Builder;
+import lombok.Getter;
+
+@OwnedBy(HarnessTeam.IDP)
+@Getter
+@Builder
+public class AppConfigDTO {
+}
