--- 101-idp-service/src/main/java/io/harness/idp/dtos/AppConfigDTO.java
@@ -1,12 +1,0 @@
-package io.harness.idp.dtos;
-
-import io.harness.annotations.dev.HarnessTeam;
-import io.harness.annotations.dev.OwnedBy;
-import lombok.Builder;
-import lombok.Getter;
-
-@OwnedBy(HarnessTeam.IDP)
-@Getter
-@Builder
-public class AppConfigDTO {
-}
