--- 101-idp-service/src/main/java/io/harness/idp/repositories/appconfig/AppConfigRepositoryCustom.java
@@ -1,4 +1,0 @@
-package io.harness.idp.repositories.appconfig;
-
-public interface AppConfigRepositoryCustom {
-}
