--- 101-idp-service/src/main/java/io/harness/idp/repositories/environmentvariable/EnvironmentVariableRepositoryCustom.java
@@ -1,4 +1,0 @@
-package io.harness.idp.repositories.environmentvariable;
-
-public interface EnvironmentVariableRepositoryCustom {
-}
