--- 101-idp-service/src/main/java/io/harness/idp/repositories/environmentvariable/EnvironmentVariableRepositoryCustomImpl.java
@@ -1,11 +1,0 @@
-package io.harness.idp.repositories.environmentvariable;
-
-import com.google.inject.Inject;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import org.springframework.data.mongodb.core.MongoTemplate;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE, onConstructor = @__({ @Inject }))
-public class EnvironmentVariableRepositoryCustomImpl implements EnvironmentVariableRepositoryCustom {
-    private MongoTemplate mongoTemplate;
-}
