--- 101-idp-service/src/main/java/io/harness/idp/serializer/kryo/IDPServiceKryoRegistrar.java
@@ -1,21 +1,0 @@
-/*
- * Copyright 2021 Harness Inc. All rights reserved.
- * Use of this source code is governed by the PolyForm Shield 1.0.0 license
- * that can be found in the licenses directory at the root of this repository, also available at
- * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
- */
-
-package io.harness.idp.serializer.kryo;
-
-import com.esotericsoftware.kryo.Kryo;
-import io.harness.annotations.dev.HarnessTeam;
-import io.harness.annotations.dev.OwnedBy;
-import io.harness.serializer.KryoRegistrar;
-
-@OwnedBy(HarnessTeam.IDP)
-public class IDPServiceKryoRegistrar implements KryoRegistrar {
-
-  @Override
-  public void register(Kryo kryo) {
-  }
-}
