--- 101-idp-service/src/main/java/io/harness/idp/serializer/morphia/IDPServiceMorphiaRegistrar.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2021 Harness Inc. All rights reserved.
- * Use of this source code is governed by the PolyForm Shield 1.0.0 license
- * that can be found in the licenses directory at the root of this repository, also available at
- * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
- */
-
-package io.harness.idp.serializer.morphia;
-
-import io.harness.annotations.dev.HarnessTeam;
-import io.harness.annotations.dev.OwnedBy;
-import io.harness.idp.entities.AppConfig;
-import io.harness.idp.entities.EnvironmentVariable;
-import io.harness.morphia.MorphiaRegistrar;
-import io.harness.morphia.MorphiaRegistrarHelperPut;
-
-import java.util.Set;
-
-@OwnedBy(HarnessTeam.IDP)
-public class IDPServiceMorphiaRegistrar implements MorphiaRegistrar {
-  @Override
-  public void registerClasses(Set<Class> set) {
-    set.add(AppConfig.class);
-    set.add(EnvironmentVariable.class);
-  }
-
-  @Override
-  public void registerImplementationClasses(MorphiaRegistrarHelperPut h, MorphiaRegistrarHelperPut w) {
-    // Nothing to register
-  }
-}
