--- 101-idp-service/src/main/java/io/harness/idp/service/appconfig/AppConfigService.java
@@ -1,4 +1,0 @@
-package io.harness.idp.service.appconfig;
-
-public interface AppConfigService {
-}
