--- 101-idp-service/src/main/java/io/harness/idp/service/configmanager/ConfigManager.java
@@ -1,4 +1,0 @@
-package io.harness.idp.service.configmanager;
-
-public interface ConfigManager {
-}
