--- 101-idp-service/src/main/java/io/harness/idp/service/configmanager/ConfigManagerImpl.java
@@ -1,4 +1,0 @@
-package io.harness.idp.service.configmanager;
-
-public class ConfigManagerImpl implements ConfigManager {
-}
