--- 101-idp-service/src/main/java/io/harness/idp/service/environmentvariable/EnvironmentVariableService.java
@@ -1,9 +1,0 @@
-package io.harness.idp.service.environmentvariable;
-
-import io.harness.idp.dtos.EnvironmentVariableDTO;
-
-import java.util.Optional;
-
-public interface EnvironmentVariableService {
-    Optional<EnvironmentVariableDTO> findByEnvName(String envName, String accountIdentifier);
-}
