--- 101-idp-service/src/main/java/io/harness/idp/service/secretmanager/SecretManager.java
@@ -1,5 +1,0 @@
-package io.harness.idp.service.secretmanager;
-
-public interface SecretManager {
-    String getSecretIdByEnvName(String envName, String accountIdentifier);
-}
