+++ 101-idp-service/src/main/java/io/harness/mappers/AppConfigMapper.java
@@ -1,0 +1,19 @@
+package io.harness.mappers;
+
+import io.harness.annotations.dev.HarnessTeam;
+import io.harness.annotations.dev.OwnedBy;
+import io.harness.dtos.AppConfigDTO;
+import io.harness.entities.AppConfig;
+import lombok.experimental.UtilityClass;
+
+@OwnedBy(HarnessTeam.IDP)
+@UtilityClass
+public class AppConfigMapper {
+    public AppConfigDTO toDTO(AppConfig appConfig) {
+        return AppConfigDTO.builder().build();
+    }
+
+    public AppConfig fromDTO(AppConfigDTO appConfigDTO) {
+        return AppConfig.builder().build();
+    }
+}
