+++ 101-idp-service/src/main/java/io/harness/repositories/appconfig/AppConfigRepositoryCustom.java
@@ -1,0 +1,4 @@
+package io.harness.repositories.appconfig;
+
+public interface AppConfigRepositoryCustom {
+}
