+++ 101-idp-service/src/main/java/io/harness/repositories/environmentvariable/EnvironmentVariableRepositoryCustom.java
@@ -1,0 +1,4 @@
+package io.harness.repositories.environmentvariable;
+
+public interface EnvironmentVariableRepositoryCustom {
+}
