+++ 101-idp-service/src/main/java/io/harness/service/appconfig/AppConfigService.java
@@ -1,0 +1,4 @@
+package io.harness.service.appconfig;
+
+public interface AppConfigService {
+}
