+++ 101-idp-service/src/main/java/io/harness/service/configmanager/ConfigManager.java
@@ -1,0 +1,4 @@
+package io.harness.service.configmanager;
+
+public interface ConfigManager {
+}
