+++ 101-idp-service/src/main/java/io/harness/service/environmentvariable/EnvironmentVariableService.java
@@ -1,0 +1,9 @@
+package io.harness.service.environmentvariable;
+
+import io.harness.dtos.EnvironmentVariableDTO;
+
+import java.util.Optional;
+
+public interface EnvironmentVariableService {
+    Optional<EnvironmentVariableDTO> findByEnvName(String envName, String accountIdentifier);
+}
