+++ 101-idp-service/src/main/java/io/harness/service/secretmanager/SecretManager.java
@@ -1,0 +1,5 @@
+package io.harness.service.secretmanager;
+
+public interface SecretManager {
+    String getSecretIdByEnvName(String envName, String accountIdentifier);
+}
